IIA-2137 created new class PatternSemanticListCell, no longer using anonymous class in PatternNavEntryController
